DatDoAnBE/requirement.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-enumfield==3.1
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jwcrypto==1.5.4
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
